share nativeChangeFieldValue between components
