+++ lib/Datepicker/Datepicker.js
@@ -8,6 +8,7 @@ import contains from 'dom-helpers/query/contains';
 import uniqueId from 'lodash/uniqueId';
 import pick from 'lodash/pick';
 import RootCloseWrapper from '../../util/RootCloseWrapper';
+import nativeChangeField from '../../util/nativeChangeFieldValue';
 import formField from '../FormField';
 import Popper, { AVAILABLE_PLACEMENTS } from '../Popper';
 import IconButton from '../IconButton';
@@ -22,22 +23,6 @@ export function getMomentLocalizedFormat(intl) {
   const format = moment.localeData()._longDateFormat.L;
   return format;
 }
-
-// Programmatically change a field value and trigger the onChange event so that external state can be updated...
-const nativeChangeField = (inputRef, focus, value = '', triggerChange = true) => {
-  if (inputRef.current) {
-    const nativeInputValueSetter = Object.getOwnPropertyDescriptor(window.HTMLInputElement.prototype, 'value').set;
-    nativeInputValueSetter.call(inputRef.current, value);
-
-    if (triggerChange) {
-      const ev = new Event('change', { bubbles: true });
-      inputRef.current.dispatchEvent(ev);
-      if (focus) {
-        inputRef.current.focus();
-      }
-    }
-  }
-};
 
 // Returns a localized format.
 export const getLocalFormat = ({ intl }) => {
