+++ lib/TextField/TextField.js
@@ -6,6 +6,7 @@ import classNames from 'classnames';
 
 import formField from '../FormField';
 import Label from '../Label';
+import nativeChangeField from '../../util/nativeChangeFieldValue';
 import TextFieldIcon from './TextFieldIcon';
 import css from './TextField.css';
 import formStyles from '../sharedStyles/form.css';
@@ -164,6 +165,7 @@ class TextField extends Component {
     hasClearIcon: true,
     type: 'text',
     validationEnabled: true,
+    autoComplete: 'off',
     value: ''
   };
 
@@ -276,12 +278,7 @@ Use the "inputRef" prop instead to obtain a ref to the input.`);
 
     // Clear value on input natively, dispatch an event to be picked up by handleChange, and focus on input again
     if (this.input.current) {
-      const nativeInputValueSetter = Object.getOwnPropertyDescriptor(window.HTMLInputElement.prototype, 'value').set;
-      nativeInputValueSetter.call(this.input.current, '');
-
-      const ev = new Event('change', { bubbles: true });
-      this.input.current.dispatchEvent(ev);
-      this.input.current.focus();
+      nativeChangeField(this.input, true, '');
     }
   }
